N21-2103 add missing condition for empty licenses to delete

+++ apps/server/src/infra/sync/media-licenses/service/vidis-sync.service.ts
@@ -92,11 +92,13 @@ export class VidisSyncService {
 			[]
 		);
 
-		await this.mediaSchoolLicenseService.deleteSchoolLicenses(licensesToDelete);
-
-		const licensesAfterDelete: MediaSchoolLicense[] = existingLicenses.filter(
-			(existingLicense: MediaSchoolLicense) => !licensesToDelete.includes(existingLicense)
-		);
+		let licensesAfterDelete: MediaSchoolLicense[] = [...existingLicenses];
+		if (licensesToDelete.length) {
+			await this.mediaSchoolLicenseService.deleteSchoolLicenses(licensesToDelete);
+			licensesAfterDelete = existingLicenses.filter(
+				(existingLicense: MediaSchoolLicense) => !licensesToDelete.includes(existingLicense)
+			);
+		}
 
 		return licensesAfterDelete;
 	}
